% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capitalize_first.R
\name{capitalize_first}
\alias{capitalize_first}
\title{Title}
\usage{
capitalize_first(s)
}
\arguments{
\item{s}{a string}
}
\value{
the string with the first letter capitalized and everything else lowercase
}
\description{
Title
}
\examples{
capitalize_first('deLIGHT')
}
