% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/city_fips.R
\name{city_fips}
\alias{city_fips}
\title{Title}
\usage{
city_fips(FIPS)
}
\arguments{
\item{FIPS}{FIPS codes of cities}
}
\value{
FIPS codes of counties associated with cities that are in the NYT covid dataset
}
\description{
Title
}
\examples{
city_fips('36998')

}
