% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete_sampler.R
\name{complete_sampler}
\alias{complete_sampler}
\title{Title}
\usage{
complete_sampler(covid_df, nc = 1)
}
\arguments{
\item{covid_df}{dataframe of covid cases from NYT}

\item{nc}{number of cores}
}
\value{
a dataframe of sampled points with their dates
}
\description{
Title
}
\examples{
complete_sampler(covid_march_2020)

}
