% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contiguous.R
\name{contiguous}
\alias{contiguous}
\title{Title}
\usage{
contiguous(df)
}
\arguments{
\item{df}{dataframe with STATEFP as a column}
}
\value{
a dataframe with just information from the contiguous US
}
\description{
Title
}
\examples{
df_contiguous <- contiguous(df)

}
