% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contiguous_v2.R
\name{contiguous_v2}
\alias{contiguous_v2}
\title{Title}
\usage{
contiguous_v2(vec)
}
\arguments{
\item{vec}{vector of STATEFP values}
}
\value{
return a vector of boolean values that describes if each entry supplied in \code{vec} is a contiguous state.
}
\description{
Title
}
\examples{
df_contiguous <- contiguous_v2(c('01', '02', '03', '04', '05'))

}
