% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_state_names.R
\name{convert_state_names}
\alias{convert_state_names}
\title{Title}
\usage{
convert_state_names(states)
}
\arguments{
\item{states}{string of states}
}
\value{
a string of properly formatted states
}
\description{
Title
}
\examples{
convert_state_names(c('nj', 'pa', 'NY'))
}
