% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_state_names_FIPS.R
\name{convert_state_names_FIPS}
\alias{convert_state_names_FIPS}
\title{Title}
\usage{
convert_state_names_FIPS(states)
}
\arguments{
\item{states}{a vector containing state names, presumably poorly formatted in the following ways:
- their two letter abbreviation
- poor capitalization of the state names}
}
\value{
vector of FIPS for the respective states
}
\description{
Title
}
\examples{
convert_state_names_FIPS(c('nj', 'pa', 'NY'))

}
