% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{covid_march_2020}
\alias{covid_march_2020}
\title{County-level covid cases in the month of March, 2020}
\format{
\subsection{\code{covid_march_2020}}{

A data frame with 13556 rows and 12 columns
}
}
\source{
\url{https://github.com/nytimes/covid-19-data/blob/master/us-counties-2020.csv}
}
\usage{
covid_march_2020
}
\description{
County-level covid cases in the month of March, 2020
}
\keyword{datasets}
