% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{df_bg}
\alias{df_bg}
\title{Shape files for each state}
\format{
\subsection{\code{df_bg}}{

A data frame with 56 rows and 10 columns:
\describe{
\item{country}{Country name}
\item{iso2, iso3}{2 & 3 letter ISO country codes}
\item{year}{Year}
...
}
}
}
\source{
\url{https://www.who.int/teams/global-tuberculosis-programme/data}
}
\usage{
df_bg
}
\description{
Shape files for each state
}
\keyword{datasets}
