% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_county_fips.R
\name{get_county_fips}
\alias{get_county_fips}
\title{Obtaining FIPS codes for counties}
\usage{
get_county_fips(states)
}
\arguments{
\item{states}{vector of states we want to know the county names of}
}
\value{
a vector of countys of the states we selected and their FIPS
}
\description{
Obtaining FIPS codes for counties
}
\examples{
get_county_fips(c('CA', 'oregon', 'nj'))
}
