% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_county_names.R
\name{get_county_names}
\alias{get_county_names}
\title{Obtaining a list of counties for a state or several states.}
\usage{
get_county_names(states)
}
\arguments{
\item{states}{vector of states we want to know the county names of}
}
\value{
a vector of countys of the states we selected
}
\description{
Obtaining a list of counties for a state or several states.
}
\examples{
get_county_names(c('CA', 'oregon', 'nj'))
}
