% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_zenodo.R
\name{load_zenodo}
\alias{load_zenodo}
\title{Title}
\usage{
load_zenodo(data, ID_number = 12981327)
}
\arguments{
\item{data}{name of the rda file to download from zenodo}

\item{ID_number}{number corresponding to the zenodo database. It will vary by the dataset version}
}
\value{
the dataset
}
\description{
Title
}
\examples{
load_zenodo('df_state.rda', 12981327)
}
