% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{los_angeles}
\alias{los_angeles}
\title{population for each block group}
\format{
\subsection{\code{los_angeles}}{

A data frame with 2495 rows and 10 columns:
}
}
\source{
\url{https://www2.census.gov/geo/tiger/GENZ2020/shp/cb_2020_us_all_500k.zip}
}
\usage{
los_angeles
}
\description{
population for each block group
}
\keyword{datasets}
