% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_shape_pop.R
\name{merge_shape_pop}
\alias{merge_shape_pop}
\title{Title}
\usage{
merge_shape_pop(shape_df)
}
\arguments{
\item{shape_df}{}
}
\value{
dataframe with pop column
}
\description{
Title
}
\examples{
merge_shape_pop(tracts('CA'))
}
