% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num_regions.R
\name{num_regions}
\alias{num_regions}
\title{Title}
\usage{
num_regions(area_fips, level)
}
\arguments{
\item{area_fips}{vector of either 2-digit or 5-digit FIPS codes for states or counties respectively}

\item{level}{integer indicating the level of granularity (2 = counties, 3 = tracts, 4 = block groups)}
}
\value{
number of regions
}
\description{
Title
}
\examples{
num_regions('06', 3)
}
