% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obtain_shapes_parallel.R
\name{obtain_shapes_parallel}
\alias{obtain_shapes_parallel}
\title{Title}
\usage{
obtain_shapes_parallel(temp_FIPS)
}
\arguments{
\item{FIPS}{}
}
\value{
temp
}
\description{
Title
}
\examples{
obtain_shapes_parallel('00000')
}
