% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obtain_shapes_parallel_v2.R
\name{obtain_shapes_parallel_v2}
\alias{obtain_shapes_parallel_v2}
\title{Title}
\usage{
obtain_shapes_parallel_v2(state_FIPS)
}
\arguments{
\item{FIPS}{}
}
\value{
temp
}
\description{
Title
}
\examples{
obtain_shapes_parallel_v2('00000')
}
