% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obtain_shapes_v2.R
\name{obtain_shapes_v2}
\alias{obtain_shapes_v2}
\title{Title}
\usage{
obtain_shapes_v2(relevant_states)
}
\arguments{
\item{FIPS}{}
}
\value{
temp
}
\description{
Title
}
\examples{
obtain_shapes_v2('36')
}
