% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obtain_specific_tracts.R
\name{obtain_specific_tracts}
\alias{obtain_specific_tracts}
\title{Title}
\usage{
obtain_specific_tracts(temp_FIPS, df)
}
\arguments{
\item{temp_FIPS}{temp}

\item{df}{temp}
}
\value{
temp
}
\description{
Title
}
\examples{
obtain_specific
}
