% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/points_to_sf.R
\name{points_to_sf}
\alias{points_to_sf}
\title{Title}
\usage{
points_to_sf(pts, nc = 1L)
}
\arguments{
\item{pts}{a 2-column dataframe with the number of points in the first column
and the respective multi-polygon to be sampled from in the
second column}

\item{nc}{}
}
\value{
a set of points in sf form, ready to plot
}
\description{
Title
}
