% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/points_to_sf_v2.R
\name{points_to_sf_v2}
\alias{points_to_sf_v2}
\title{Title}
\usage{
points_to_sf_v2(pop_vec, geo_vec, dat_vec, nc = 1L)
}
\arguments{
\item{nc}{}

\item{pts}{v2: a 3 column dataframe (sampled_pop, geometry, dates)
sampled_pop\link{1} is a list, we must unlist() it
geometry\link{1} is a list, we must unlist() it and convert it into a matrix
dates\link{1} is a list, we must unlist() it, unname() it, and convert it from 5 digit integer to date using unix}
}
\value{
a set of points in sf form, ready to plot
}
\description{
Title
}
