% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_data.R
\name{prep_data}
\alias{prep_data}
\title{Title}
\usage{
prep_data(shape_names, level)
}
\arguments{
\item{shape_names}{string of county names or state names}

\item{level}{level of sampling granularity}
}
\value{
a dataframe of the relevant regions
}
\description{
Title
}
\examples{
prep_data('california', 3)

}
