% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reassign_county_cases.R
\name{reassign_county_cases}
\alias{reassign_county_cases}
\title{Title}
\usage{
reassign_county_cases(cases_FIPS, temp_df, used_metros)
}
\arguments{
\item{cases_FIPS}{}
}
\value{
temp
}
\description{
Title
}
\examples{
reassign_county_cases(covid_march_2020)
}
