% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reassign_county_cases_v2_troubleshoot.R
\name{reassign_county_cases_v2_troubleshoot}
\alias{reassign_county_cases_v2_troubleshoot}
\title{Title}
\usage{
reassign_county_cases_v2_troubleshoot(temp_FIPS, N, all_dates)
}
\arguments{
\item{temp_FIPS}{a string for the FIPS}

\item{N}{a number for number of FIPS}

\item{all_dates}{a vector of date objects}
}
\value{
child_df2, a dataframe of 3 columns and N rows. The three columns describe FIPS,
}
\description{
Title
}
\examples{
reassign_county_cases_v2_troubleshoot(covid_march_2020)
}
