% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samp.R
\name{samp}
\alias{samp}
\title{Title}
\usage{
samp(freq, geo)
}
\arguments{
\item{freq}{integer denoting how many samples}

\item{geo}{sf object denoting the area to be randomly sampled}
}
\value{
a list of sampled points
}
\description{
Title
}
\examples{
samp(3, los_angeles$geometry[2])
}
