% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_by_population.R
\name{sample_by_population}
\alias{sample_by_population}
\title{Sampling from a collection shapes, each with a specific population}
\usage{
sample_by_population(N, df)
}
\arguments{
\item{N}{number of samples}

\item{df}{a dataframe with population (pop) and shape (geometry) columns}
}
\value{
a list of sampled locations
}
\description{
Sampling from a collection shapes, each with a specific population
}
\examples{
df <- data.frame(pop = c(1,1,1), geometry = shapes)
sample_by_population(1000, df)
}
