% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_by_population_V2.R
\name{sample_by_population_V2}
\alias{sample_by_population_V2}
\title{Sampling from a collection shapes, each with a specific population}
\usage{
sample_by_population_V2(N, area_fips, level)
}
\arguments{
\item{N}{number of samples}

\item{area_fips}{a dataframe with population (pop) and shape (geometry) columns}

\item{level}{granularity}
}
\value{
a list of sampled locations
}
\description{
Sampling from a collection shapes, each with a specific population
}
\examples{
df <- data.frame(pop = c(1,1,1), geometry = shapes)
sample_by_population(1000, df)
}
