% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampled_region_outline.R
\name{sampled_region_outline}
\alias{sampled_region_outline}
\title{Title}
\usage{
sampled_region_outline(area_fips, level)
}
\arguments{
\item{area_fips}{vector of fips}

\item{level}{the level of granularity}
}
\value{
shape
}
\description{
Title
}
\examples{
sampled_region_outline(c('06', '14'))
}
