% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampled_region_outline_v2.R
\name{sampled_region_outline_v2}
\alias{sampled_region_outline_v2}
\title{Title}
\usage{
sampled_region_outline_v2(area_fips, level)
}
\arguments{
\item{area_fips}{vector of fips}

\item{level}{the level of granularity depicted in the plot (0 = country, 1 = state, 2 = county)}
}
\value{
shape
}
\description{
Title
}
\examples{
sampled_region_outline_v2(c('06', '32'), 2)

}
