% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_metropolis_FIPS_v2.R
\name{split_metropolis_FIPS_v2}
\alias{split_metropolis_FIPS_v2}
\title{Title}
\usage{
split_metropolis_FIPS_v2(cases_FIPS)
}
\arguments{
\item{cases_FIPS}{dataframe of FIPS with their cases}
}
\value{
dataframe
}
\description{
Title
}
\examples{
split_metropolis_FIPS(covid_march_2020)
}
