% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{state_pop}
\alias{state_pop}
\title{population for each state}
\format{
\subsection{\code{state_pop}}{

A data frame with 56 rows and 10 columns:
\describe{
\item{country}{Country name}
\item{iso2, iso3}{2 & 3 letter ISO country codes}
\item{year}{Year}
...
}
}
}
\source{
\url{https://www.who.int/teams/global-tuberculosis-programme/data}
}
\usage{
state_pop
}
\description{
population for each state
}
\keyword{datasets}
