% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{statefp}
\alias{statefp}
\title{United States FIPS numbers for each state}
\format{
\subsection{\code{statefp}}{

A data frame with 56 rows and 2 columns:
\describe{
\item{country}{Country name}
\item{iso2, iso3}{2 & 3 letter ISO country codes}
\item{year}{Year}
...
}
}
}
\source{
\url{https://www.who.int/teams/global-tuberculosis-programme/data}
}
\usage{
statefp
}
\description{
United States FIPS numbers for each state
}
\keyword{datasets}
