% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{tract_shapes}
\alias{tract_shapes}
\title{Shape files for each tract}
\format{
\subsection{\code{tract_shapes}}{

A data frame with 56 rows and 10 columns:
\describe{
\item{country}{Country name}
\item{iso2, iso3}{2 & 3 letter ISO country codes}
\item{year}{Year}
...
}
}
}
\source{
\url{https://www.who.int/teams/global-tuberculosis-programme/data}
}
\usage{
tract_shapes
}
\description{
Shape files for each tract
}
\keyword{datasets}
