% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_state_names.R
\name{convert_state_names}
\alias{convert_state_names}
\title{Title}
\usage{
convert_state_names(states)
}
\arguments{
\item{states}{a vector containing state names, presumably poorly formatted in the following ways:
- their two letter abbreviation
- poor capitalization of the state names}
}
\value{
a string of states formatted by their full names with proper capitalization
}
\description{
Title
}
\examples{
convert_state_names(c('nj', 'pa', 'NY'))
}
