% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_state_names_FIPS.R
\name{convert_state_names_FIPS}
\alias{convert_state_names_FIPS}
\title{Title}
\usage{
convert_state_names_FIPS(states)
}
\arguments{
\item{states}{vector of states}
}
\value{
vector of FIPS for the states
}
\description{
Title
}
\examples{
...
}
