% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler_helper.R
\name{sampler_helper}
\alias{sampler_helper}
\title{sampler :D}
\usage{
sampler_helper(N, df)
}
\arguments{
\item{N}{number of samples}

\item{df}{a dataframe with population (pop) and shape (geometry) columns}
}
\value{
sampled points
}
\description{
sampler :D
}
\examples{
sample_by_population(N, df)
}
