% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler_helper.R
\name{sampler_helper}
\alias{sampler_helper}
\title{Helper function to uniformly draw points in a shape proportional to their population}
\usage{
sampler_helper(df, N)
}
\arguments{
\item{df}{a dataframe with population (pop) and shape (geometry) columns}

\item{N}{number of samples}
}
\value{
a list of sampled points
}
\description{
Helper function to uniformly draw points in a shape proportional to their population
}
\examples{
sampler_helper(los_angeles, 10)
}
