% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler_helper_parallel.R
\name{sampler_helper_parallel}
\alias{sampler_helper_parallel}
\title{Helper function to uniformly draw points in a shape proportional to their population}
\usage{
sampler_helper_parallel(N, df)
}
\arguments{
\item{N}{number of samples}

\item{df}{a dataframe with population (pop) and shape (geometry) columns}
}
\value{
a list of sampled points
}
\description{
Helper function to uniformly draw points in a shape proportional to their population
}
\examples{
sampler_helper_parallel(10, los_angeles)
}
